requirements.txt
djangorestframework==3.16.1
djangorestframework-simplejwt==4.7.2
Django==4.2
...
